fix a couple of warnings.
many many more left

+++ src/graph.hpp
@@ -663,7 +663,7 @@ template <class adjacency_struct>
 void graph<adjacency_struct>::add_edge(int u, int v)
 {
 
-    if (matrix.size() <= v) {
+    if (matrix.size() <= static_cast<size_t>(v)) {
         std::cout << "ERROR: try to add edge " << u << "-" << v << " / "
                   << nodes.size() << " " << nodeset.size() << " "
                   << matrix.size() << std::endl;
